[test]rebuild code
