+++ test/e2e/E2eTest/ls_comprehensive_test.go
@@ -2,12 +2,13 @@ package E2eTest
 
 import (
 	"context"
-	k8serror "k8s.io/apimachinery/pkg/api/errors"
-	"time"
+	"github.com/hwameistor/hwameistor/test/e2e/utils"
 
-	clientset "github.com/hwameistor/hwameistor/pkg/apis/client/clientset/versioned/scheme"
 	v1alpha1 "github.com/hwameistor/hwameistor/pkg/apis/hwameistor/v1alpha1"
 	"github.com/hwameistor/hwameistor/test/e2e/framework"
+
+	"time"
+
 	"github.com/onsi/ginkgo/v2"
 	"github.com/onsi/gomega"
 	"github.com/sirupsen/logrus"
@@ -15,6 +16,7 @@ import (
 	apiv1 "k8s.io/api/core/v1"
 	v1 "k8s.io/api/core/v1"
 	storagev1 "k8s.io/api/storage/v1"
+	k8serror "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/labels"
@@ -24,16 +26,246 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client/config"
 )
 
-var _ = ginkgo.Describe("test localstorage volume ", ginkgo.Label("periodCheck"), func() {
+var _ = ginkgo.Describe("comprehensive test", func() {
 
-	f := framework.NewDefaultFramework(clientset.AddToScheme)
+	f := framework.NewDefaultFramework(v1alpha1.AddToScheme)
 	client := f.GetClient()
 	ctx := context.TODO()
 	ginkgo.It("Configure the base environment", func() {
-		result := configureEnvironment(ctx)
+		result := utils.ConfigureEnvironment(ctx)
 		gomega.Expect(result).To(gomega.BeNil())
-		createLdc(ctx)
 
+	})
+	ginkgo.Context("create a StorageClass", func() {
+		ginkgo.It("create a sc", func() {
+			//create sc
+			deleteObj := apiv1.PersistentVolumeReclaimDelete
+			waitForFirstConsumerObj := storagev1.VolumeBindingWaitForFirstConsumer
+			examplesc := &storagev1.StorageClass{
+				ObjectMeta: metav1.ObjectMeta{
+					Name: "local-storage-hdd-fs",
+				},
+				Provisioner: "disk.hwameistor.io",
+				Parameters: map[string]string{
+					"diskType": "HDD",
+				},
+				ReclaimPolicy:        &deleteObj,
+				AllowVolumeExpansion: utils.BoolPter(false),
+				VolumeBindingMode:    &waitForFirstConsumerObj,
+			}
+			err := client.Create(ctx, examplesc)
+			if err != nil {
+				logrus.Printf("Create SC failed ：%+v ", err)
+				f.ExpectNoError(err)
+			}
+		})
+	})
+	ginkgo.Context("create a PVC", func() {
+		ginkgo.It("create PVC", func() {
+			//create PVC
+			storageClassName := "local-storage-hdd-fs"
+			examplePvc := &apiv1.PersistentVolumeClaim{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "pvc-fs",
+					Namespace: "default",
+				},
+				Spec: apiv1.PersistentVolumeClaimSpec{
+					AccessModes:      []apiv1.PersistentVolumeAccessMode{apiv1.ReadWriteOnce},
+					StorageClassName: &storageClassName,
+					Resources: apiv1.ResourceRequirements{
+						Requests: apiv1.ResourceList{
+							apiv1.ResourceStorage: resource.MustParse("1Gi"),
+						},
+					},
+				},
+			}
+			err := client.Create(ctx, examplePvc)
+			if err != nil {
+				logrus.Printf("Create PVC failed ：%+v ", err)
+				f.ExpectNoError(err)
+			}
+
+			gomega.Expect(err).To(gomega.BeNil())
+		})
+
+	})
+	ginkgo.Context("create a deployment", func() {
+
+		ginkgo.It("create deployment", func() {
+			//create deployment
+			exampleDeployment := &appsv1.Deployment{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      utils.DeploymentName,
+					Namespace: "default",
+				},
+				Spec: appsv1.DeploymentSpec{
+					Replicas: utils.Int32Ptr(1),
+					Selector: &metav1.LabelSelector{
+						MatchLabels: map[string]string{
+							"app": "demo",
+						},
+					},
+					Template: apiv1.PodTemplateSpec{
+						ObjectMeta: metav1.ObjectMeta{
+							Labels: map[string]string{
+								"app": "demo",
+							},
+						},
+						Spec: apiv1.PodSpec{
+							SchedulerName: "hwameistor-scheduler",
+							Containers: []apiv1.Container{
+								{
+									Name:  "web",
+									Image: "ghcr.m.daocloud.io/daocloud/dao-2048:v1.2.0",
+									Ports: []apiv1.ContainerPort{
+										{
+											Name:          "http",
+											Protocol:      apiv1.ProtocolTCP,
+											ContainerPort: 80,
+										},
+									},
+									VolumeMounts: []apiv1.VolumeMount{
+										{
+											Name:      "2048-volume-lvm",
+											MountPath: "/data",
+										},
+									},
+								},
+							},
+							Volumes: []apiv1.Volume{
+								{
+									Name: "2048-volume-lvm",
+									VolumeSource: apiv1.VolumeSource{
+										PersistentVolumeClaim: &apiv1.PersistentVolumeClaimVolumeSource{
+											ClaimName: "pvc-fs",
+										},
+									},
+								},
+							},
+						},
+					},
+				},
+			}
+			err := client.Create(ctx, exampleDeployment)
+			if err != nil {
+				logrus.Printf("%+v ", err)
+				f.ExpectNoError(err)
+			}
+			gomega.Expect(err).To(gomega.BeNil())
+		})
+		ginkgo.It("PVC STATUS should be Bound", func() {
+			pvc := &apiv1.PersistentVolumeClaim{}
+			pvcKey := k8sclient.ObjectKey{
+				Name:      "pvc-fs",
+				Namespace: "default",
+			}
+			err := client.Get(ctx, pvcKey, pvc)
+			if err != nil {
+				logrus.Printf("%+v ", err)
+				f.ExpectNoError(err)
+			}
+
+			logrus.Infof("Waiting for the PVC to be bound")
+			err = wait.PollImmediate(3*time.Second, 3*time.Minute, func() (done bool, err error) {
+				if err = client.Get(ctx, pvcKey, pvc); pvc.Status.Phase != apiv1.ClaimBound {
+					return false, nil
+				}
+				return true, nil
+			})
+			if err != nil {
+				logrus.Infof("PVC binding timeout")
+				logrus.Error(err)
+			}
+			gomega.Expect(err).To(gomega.BeNil())
+		})
+		ginkgo.It("deploy STATUS should be AVAILABLE", func() {
+			deployment := &appsv1.Deployment{}
+			deployKey := k8sclient.ObjectKey{
+				Name:      utils.DeploymentName,
+				Namespace: "default",
+			}
+			err := client.Get(ctx, deployKey, deployment)
+			if err != nil {
+				logrus.Printf("%+v ", err)
+				f.ExpectNoError(err)
+			}
+			logrus.Infof("waiting for the deployment to be ready ")
+			err = wait.PollImmediate(3*time.Second, 3*time.Minute, func() (done bool, err error) {
+				if err = client.Get(ctx, deployKey, deployment); deployment.Status.AvailableReplicas != int32(1) {
+					return false, nil
+				}
+				return true, nil
+			})
+			if err != nil {
+				logrus.Infof("deployment ready timeout")
+				logrus.Error(err)
+			}
+			gomega.Expect(err).To(gomega.BeNil())
+
+		})
+
+	})
+	ginkgo.Context("Clean up the environment", func() {
+		ginkgo.It("Delete test Deployment", func() {
+			//delete deploy
+			deployment := &appsv1.Deployment{}
+			deployKey := k8sclient.ObjectKey{
+				Name:      utils.DeploymentName,
+				Namespace: "default",
+			}
+			err := client.Get(ctx, deployKey, deployment)
+			if err != nil {
+				logrus.Error(err)
+				f.ExpectNoError(err)
+			}
+			logrus.Infof("deleting test Deployment ")
+
+			err = client.Delete(ctx, deployment)
+			if err != nil {
+				logrus.Error(err)
+				f.ExpectNoError(err)
+			}
+			err = wait.PollImmediate(3*time.Second, 3*time.Minute, func() (done bool, err error) {
+				if err := client.Get(ctx, deployKey, deployment); !k8serror.IsNotFound(err) {
+					return false, nil
+				}
+				return true, nil
+			})
+			if err != nil {
+				logrus.Error(err)
+			}
+			gomega.Expect(err).To(gomega.BeNil())
+
+		})
+		ginkgo.It("delete all pvc ", func() {
+			err := utils.DeleteAllPVC(ctx)
+			gomega.Expect(err).To(gomega.BeNil())
+		})
+		ginkgo.It("check pv", func() {
+			logrus.Printf("check pv")
+			pvList := &apiv1.PersistentVolumeList{}
+
+			err := wait.PollImmediate(3*time.Second, 3*time.Minute, func() (done bool, err error) {
+				err = client.List(ctx, pvList)
+				if err != nil {
+					logrus.Error("get pv list error", err)
+					f.ExpectNoError(err)
+				}
+				if len(pvList.Items) != 0 {
+					return false, nil
+				} else {
+					return true, nil
+				}
+			})
+			if err != nil {
+				logrus.Error(err)
+			}
+			gomega.Expect(err).To(gomega.BeNil())
+		})
+		ginkgo.It("delete all sc", func() {
+			err := utils.DeleteAllSC(ctx)
+			gomega.Expect(err).To(gomega.BeNil())
+		})
 	})
 
 	ginkgo.Context("create a StorageClass", func() {
@@ -55,7 +287,7 @@ var _ = ginkgo.Describe("test localstorage volume ", ginkgo.Label("periodCheck")
 					"csi.storage.k8s.io/fstype": "xfs",
 				},
 				ReclaimPolicy:        &deleteObj,
-				AllowVolumeExpansion: boolPter(true),
+				AllowVolumeExpansion: utils.BoolPter(true),
 				VolumeBindingMode:    &waitForFirstConsumerObj,
 			}
 			err := client.Create(ctx, examplesc)
@@ -100,11 +332,11 @@ var _ = ginkgo.Describe("test localstorage volume ", ginkgo.Label("periodCheck")
 			//create deployment
 			exampleDeployment := &appsv1.Deployment{
 				ObjectMeta: metav1.ObjectMeta{
-					Name:      DeploymentName,
+					Name:      utils.DeploymentName,
 					Namespace: "default",
 				},
 				Spec: appsv1.DeploymentSpec{
-					Replicas: int32Ptr(1),
+					Replicas: utils.Int32Ptr(1),
 					Selector: &metav1.LabelSelector{
 						MatchLabels: map[string]string{
 							"app": "demo",
@@ -117,6 +349,7 @@ var _ = ginkgo.Describe("test localstorage volume ", ginkgo.Label("periodCheck")
 							},
 						},
 						Spec: apiv1.PodSpec{
+							SchedulerName: "hwameistor-scheduler",
 							Containers: []apiv1.Container{
 								{
 									Name:  "web",
@@ -185,7 +418,7 @@ var _ = ginkgo.Describe("test localstorage volume ", ginkgo.Label("periodCheck")
 		ginkgo.It("deploy STATUS should be AVAILABLE", func() {
 			deployment := &appsv1.Deployment{}
 			deployKey := k8sclient.ObjectKey{
-				Name:      DeploymentName,
+				Name:      utils.DeploymentName,
 				Namespace: "default",
 			}
 			err := client.Get(ctx, deployKey, deployment)
@@ -209,7 +442,7 @@ var _ = ginkgo.Describe("test localstorage volume ", ginkgo.Label("periodCheck")
 		})
 
 	})
-	ginkgo.Context("Test Eviction", func() {
+	ginkgo.Context("Test the volume", func() {
 		ginkgo.It("check lvg", func() {
 			lvrList := &v1alpha1.LocalVolumeReplicaList{}
 			err := client.List(ctx, lvrList)
@@ -228,7 +461,8 @@ var _ = ginkgo.Describe("test localstorage volume ", ginkgo.Label("periodCheck")
 			}
 
 		})
-		ginkgo.It("Write test file", func() {
+		ginkgo.It("write test data", func() {
+
 			config, err := config.GetConfig()
 			if err != nil {
 				return
@@ -236,7 +470,7 @@ var _ = ginkgo.Describe("test localstorage volume ", ginkgo.Label("periodCheck")
 
 			deployment := &appsv1.Deployment{}
 			deployKey := k8sclient.ObjectKey{
-				Name:      DeploymentName,
+				Name:      utils.DeploymentName,
 				Namespace: "default",
 			}
 			err = client.Get(ctx, deployKey, deployment)
@@ -262,12 +496,12 @@ var _ = ginkgo.Describe("test localstorage volume ", ginkgo.Label("periodCheck")
 			containers := deployment.Spec.Template.Spec.Containers
 			for _, pod := range podlist.Items {
 				for _, container := range containers {
-					_, _, err := ExecInPod(config, deployment.Namespace, pod.Name, "cd /data && echo it-is-a-test >test", container.Name)
+					_, _, err := utils.ExecInPod(config, deployment.Namespace, pod.Name, "cd /data && echo it-is-a-test >test", container.Name)
 					if err != nil {
 						logrus.Printf("%+v ", err)
 						f.ExpectNoError(err)
 					}
-					output, _, err := ExecInPod(config, deployment.Namespace, pod.Name, "cd /data && cat test", container.Name)
+					output, _, err := utils.ExecInPod(config, deployment.Namespace, pod.Name, "cd /data && cat test", container.Name)
 					if err != nil {
 						logrus.Printf("%+v ", err)
 						f.ExpectNoError(err)
@@ -276,68 +510,7 @@ var _ = ginkgo.Describe("test localstorage volume ", ginkgo.Label("periodCheck")
 				}
 			}
 		})
-		ginkgo.It("Evicte volume", func() {
-			deployment := &appsv1.Deployment{}
-			deployKey := k8sclient.ObjectKey{
-				Name:      DeploymentName,
-				Namespace: "default",
-			}
-			err := client.Get(ctx, deployKey, deployment)
-			if err != nil {
-				logrus.Printf("%+v ", err)
-				f.ExpectNoError(err)
-			}
-
-			apps, err := labels.NewRequirement("app", selection.In, []string{"demo"})
-			selector := labels.NewSelector()
-			selector = selector.Add(*apps)
-			listOption := k8sclient.ListOptions{
-				LabelSelector: selector,
-			}
-			podlist := &v1.PodList{}
-			err = client.List(ctx, podlist, &listOption)
-			if err != nil {
-				logrus.Printf("%+v ", err)
-				f.ExpectNoError(err)
-			}
-
-			logrus.Infof("drain  " + podlist.Items[0].Spec.NodeName)
-			beforeEvictionNode := podlist.Items[0].Spec.NodeName
-			_ = runInLinux("kubectl taint node --all node-role.kubernetes.io/master-")
-			_ = runInLinux("kubectl drain " + podlist.Items[0].Spec.NodeName + " --ignore-daemonsets=true --delete-emptydir-data")
-
-			deployment = &appsv1.Deployment{}
-			deployKey = k8sclient.ObjectKey{
-				Name:      DeploymentName,
-				Namespace: "default",
-			}
-			err = client.Get(ctx, deployKey, deployment)
-			if err != nil {
-				logrus.Printf("%+v ", err)
-				f.ExpectNoError(err)
-			}
-			logrus.Infof("waiting for the deployment to be ready ")
-			err = wait.PollImmediate(3*time.Second, 5*time.Minute, func() (done bool, err error) {
-				if err = client.Get(ctx, deployKey, deployment); deployment.Status.AvailableReplicas != int32(1) {
-					return false, nil
-				}
-				return true, nil
-			})
-			if err != nil {
-				logrus.Infof("deployment ready timeout")
-				logrus.Error(err)
-			}
-			err = client.List(ctx, podlist, &listOption)
-			if err != nil {
-				logrus.Printf("%+v ", err)
-				f.ExpectNoError(err)
-			}
-
-			logrus.Infof("after drain " + podlist.Items[0].Spec.NodeName)
-			_ = runInLinux("kubectl uncordon " + beforeEvictionNode)
-			gomega.Expect(beforeEvictionNode).NotTo(gomega.Equal(podlist.Items[0].Spec.NodeName))
-		})
-		ginkgo.It("check test file", func() {
+		ginkgo.It("Delete test data", func() {
 			config, err := config.GetConfig()
 			if err != nil {
 				return
@@ -345,7 +518,7 @@ var _ = ginkgo.Describe("test localstorage volume ", ginkgo.Label("periodCheck")
 
 			deployment := &appsv1.Deployment{}
 			deployKey := k8sclient.ObjectKey{
-				Name:      DeploymentName,
+				Name:      utils.DeploymentName,
 				Namespace: "default",
 			}
 			err = client.Get(ctx, deployKey, deployment)
@@ -371,24 +544,27 @@ var _ = ginkgo.Describe("test localstorage volume ", ginkgo.Label("periodCheck")
 			containers := deployment.Spec.Template.Spec.Containers
 			for _, pod := range podlist.Items {
 				for _, container := range containers {
-					logrus.Printf(pod.Name)
-					output, _, err := ExecInPod(config, deployment.Namespace, pod.Name, "cd /data && cat test", container.Name)
+					_, _, err := utils.ExecInPod(config, deployment.Namespace, pod.Name, "cd /data && rm -rf test", container.Name)
 					if err != nil {
 						logrus.Printf("%+v ", err)
 						f.ExpectNoError(err)
 					}
-					gomega.Expect(output).To(gomega.Equal("it-is-a-test"))
+					output, _, err := utils.ExecInPod(config, deployment.Namespace, pod.Name, "cd /data && ls", container.Name)
+					if err != nil {
+						logrus.Printf("%+v ", err)
+						f.ExpectNoError(err)
+					}
+					gomega.Expect(output).To(gomega.Equal(""))
 				}
 			}
 		})
-
 	})
 	ginkgo.Context("Clean up the environment", func() {
 		ginkgo.It("Delete test Deployment", func() {
 			//delete deploy
 			deployment := &appsv1.Deployment{}
 			deployKey := k8sclient.ObjectKey{
-				Name:      DeploymentName,
+				Name:      utils.DeploymentName,
 				Namespace: "default",
 			}
 			err := client.Get(ctx, deployKey, deployment)
@@ -416,13 +592,18 @@ var _ = ginkgo.Describe("test localstorage volume ", ginkgo.Label("periodCheck")
 
 		})
 		ginkgo.It("delete all pvc ", func() {
-			err := deleteAllPVC(ctx)
+			err := utils.DeleteAllPVC(ctx)
 			gomega.Expect(err).To(gomega.BeNil())
 		})
 		ginkgo.It("delete all sc", func() {
-			err := deleteAllSC(ctx)
+			err := utils.DeleteAllSC(ctx)
 			gomega.Expect(err).To(gomega.BeNil())
 		})
+
+	})
+
+	ginkgo.It("Configure the base environment", func() {
+		utils.CreateLdc(ctx)
 	})
 	ginkgo.Context("create a HA-StorageClass", func() {
 		ginkgo.It("create a sc", func() {
@@ -443,7 +624,7 @@ var _ = ginkgo.Describe("test localstorage volume ", ginkgo.Label("periodCheck")
 					"csi.storage.k8s.io/fstype": "xfs",
 				},
 				ReclaimPolicy:        &deleteObj,
-				AllowVolumeExpansion: boolPter(true),
+				AllowVolumeExpansion: utils.BoolPter(true),
 				VolumeBindingMode:    &waitForFirstConsumerObj,
 			}
 			err := client.Create(ctx, examplesc)
@@ -488,11 +669,11 @@ var _ = ginkgo.Describe("test localstorage volume ", ginkgo.Label("periodCheck")
 			//create deployment
 			exampleDeployment := &appsv1.Deployment{
 				ObjectMeta: metav1.ObjectMeta{
-					Name:      HaDeploymentName,
+					Name:      utils.HaDeploymentName,
 					Namespace: "default",
 				},
 				Spec: appsv1.DeploymentSpec{
-					Replicas: int32Ptr(1),
+					Replicas: utils.Int32Ptr(1),
 					Strategy: appsv1.DeploymentStrategy{
 						Type: appsv1.RecreateDeploymentStrategyType,
 					},
@@ -508,6 +689,30 @@ var _ = ginkgo.Describe("test localstorage volume ", ginkgo.Label("periodCheck")
 							},
 						},
 						Spec: apiv1.PodSpec{
+							SchedulerName: "hwameistor-scheduler",
+							Affinity: &apiv1.Affinity{
+								NodeAffinity: &apiv1.NodeAffinity{
+									RequiredDuringSchedulingIgnoredDuringExecution: &apiv1.NodeSelector{
+										NodeSelectorTerms: []apiv1.NodeSelectorTerm{
+											{
+												[]apiv1.NodeSelectorRequirement{
+													{
+														Key:      "kubernetes.io/hostname",
+														Operator: apiv1.NodeSelectorOpIn,
+														Values: []string{
+															"k8s-node1",
+														},
+													},
+												},
+												[]apiv1.NodeSelectorRequirement{},
+											},
+										},
+									},
+									PreferredDuringSchedulingIgnoredDuringExecution: nil,
+								},
+								PodAffinity:     nil,
+								PodAntiAffinity: nil,
+							},
 							Containers: []apiv1.Container{
 								{
 									Name:  "web",
@@ -577,7 +782,7 @@ var _ = ginkgo.Describe("test localstorage volume ", ginkgo.Label("periodCheck")
 		ginkgo.It("deploy STATUS should be AVAILABLE", func() {
 			deployment := &appsv1.Deployment{}
 			deployKey := k8sclient.ObjectKey{
-				Name:      HaDeploymentName,
+				Name:      utils.HaDeploymentName,
 				Namespace: "default",
 			}
 			err := client.Get(ctx, deployKey, deployment)
@@ -600,7 +805,151 @@ var _ = ginkgo.Describe("test localstorage volume ", ginkgo.Label("periodCheck")
 		})
 
 	})
+	ginkgo.Context("test volumes", func() {
+		ginkgo.It("Write test file", func() {
+			//create a request
+			config, err := config.GetConfig()
+			if err != nil {
+				return
+			}
+
+			deployment := &appsv1.Deployment{}
+			deployKey := k8sclient.ObjectKey{
+				Name:      utils.HaDeploymentName,
+				Namespace: "default",
+			}
+			err = client.Get(ctx, deployKey, deployment)
+			if err != nil {
+				logrus.Printf("%+v ", err)
+				f.ExpectNoError(err)
+			}
+
+			apps, err := labels.NewRequirement("app", selection.In, []string{"demo"})
+			selector := labels.NewSelector()
+			selector = selector.Add(*apps)
+			listOption := k8sclient.ListOptions{
+				LabelSelector: selector,
+			}
+			podlist := &v1.PodList{}
+			err = client.List(ctx, podlist, &listOption)
+
+			if err != nil {
+				logrus.Printf("%+v ", err)
+				f.ExpectNoError(err)
+			}
+
+			containers := deployment.Spec.Template.Spec.Containers
+			for _, pod := range podlist.Items {
+				for _, container := range containers {
+					_, _, err := utils.ExecInPod(config, deployment.Namespace, pod.Name, "cd /data && echo it-is-a-test >test", container.Name)
+					if err != nil {
+						logrus.Printf("%+v ", err)
+						f.ExpectNoError(err)
+					}
+					output, _, err := utils.ExecInPod(config, deployment.Namespace, pod.Name, "cd /data && cat test", container.Name)
+					if err != nil {
+						logrus.Printf("%+v ", err)
+						f.ExpectNoError(err)
+					}
+					gomega.Expect(output).To(gomega.Equal("it-is-a-test"))
+				}
+			}
+		})
+		ginkgo.It("Delete test file", func() {
+			config, err := config.GetConfig()
+			if err != nil {
+				return
+			}
+
+			deployment := &appsv1.Deployment{}
+			deployKey := k8sclient.ObjectKey{
+				Name:      utils.HaDeploymentName,
+				Namespace: "default",
+			}
+			err = client.Get(ctx, deployKey, deployment)
+			if err != nil {
+				logrus.Printf("%+v ", err)
+				f.ExpectNoError(err)
+			}
+
+			apps, err := labels.NewRequirement("app", selection.In, []string{"demo"})
+			selector := labels.NewSelector()
+			selector = selector.Add(*apps)
+			listOption := k8sclient.ListOptions{
+				LabelSelector: selector,
+			}
+			podlist := &v1.PodList{}
+			err = client.List(ctx, podlist, &listOption)
+
+			if err != nil {
+				logrus.Printf("%+v ", err)
+				f.ExpectNoError(err)
+			}
+
+			containers := deployment.Spec.Template.Spec.Containers
+			for _, pod := range podlist.Items {
+				for _, container := range containers {
+					_, _, err := utils.ExecInPod(config, deployment.Namespace, pod.Name, "cd /data && rm -rf test", container.Name)
+					if err != nil {
+						logrus.Printf("%+v ", err)
+						f.ExpectNoError(err)
+					}
+					output, _, err := utils.ExecInPod(config, deployment.Namespace, pod.Name, "cd /data && ls", container.Name)
+					if err != nil {
+						logrus.Printf("%+v ", err)
+						f.ExpectNoError(err)
+					}
+					gomega.Expect(output).To(gomega.Equal(""))
+				}
+			}
+		})
+	})
 	ginkgo.Context("test HA-volumes", func() {
+		ginkgo.It("create a localvolumemigrate", func() {
+
+			lvrList := &v1alpha1.LocalVolumeReplicaList{}
+			err := client.List(ctx, lvrList)
+			if err != nil {
+				logrus.Printf("list lvr failed ：%+v ", err)
+			}
+			lvlist := &v1alpha1.LocalVolumeList{}
+			err = client.List(ctx, lvlist)
+			if err != nil {
+				logrus.Error("%+v ", err)
+				f.ExpectNoError(err)
+			}
+
+			lvname := lvlist.Items[0].Name
+			for _, lvr := range lvrList.Items {
+				if lvr.Spec.NodeName == "k8s-master" {
+					exlvm := &v1alpha1.LocalVolumeMigrate{
+						ObjectMeta: metav1.ObjectMeta{
+							Name:      "localvolumegroupmigrate-1",
+							Namespace: "default",
+						},
+						Spec: v1alpha1.LocalVolumeMigrateSpec{
+							TargetNodesSuggested: []string{"k8s-node2"},
+							SourceNode:           "k8s-master",
+							VolumeName:           lvname,
+							MigrateAllVols:       true,
+						},
+					}
+
+					err = client.Create(ctx, exlvm)
+					logrus.Infof("create lvm")
+					if err != nil {
+						logrus.Printf("Create lvgm failed ：%+v ", err)
+						f.ExpectNoError(err)
+					}
+					logrus.Infof("wait 3 minutes for migrate lv")
+					time.Sleep(3 * time.Minute)
+					break
+
+				}
+
+			}
+
+		})
 
 		ginkgo.It("Write test file", func() {
 			config, err := config.GetConfig()
@@ -610,7 +959,7 @@ var _ = ginkgo.Describe("test localstorage volume ", ginkgo.Label("periodCheck")
 
 			deployment := &appsv1.Deployment{}
 			deployKey := k8sclient.ObjectKey{
-				Name:      HaDeploymentName,
+				Name:      utils.HaDeploymentName,
 				Namespace: "default",
 			}
 			err = client.Get(ctx, deployKey, deployment)
@@ -636,12 +985,12 @@ var _ = ginkgo.Describe("test localstorage volume ", ginkgo.Label("periodCheck")
 			containers := deployment.Spec.Template.Spec.Containers
 			for _, pod := range podlist.Items {
 				for _, container := range containers {
-					_, _, err := ExecInPod(config, deployment.Namespace, pod.Name, "cd /data && echo it-is-a-test >test", container.Name)
+					_, _, err := utils.ExecInPod(config, deployment.Namespace, pod.Name, "cd /data && echo it-is-a-test >test", container.Name)
 					if err != nil {
 						logrus.Printf("%+v ", err)
 						f.ExpectNoError(err)
 					}
-					output, _, err := ExecInPod(config, deployment.Namespace, pod.Name, "cd /data && cat test", container.Name)
+					output, _, err := utils.ExecInPod(config, deployment.Namespace, pod.Name, "cd /data && cat test", container.Name)
 					if err != nil {
 						logrus.Printf("%+v ", err)
 						f.ExpectNoError(err)
@@ -650,48 +999,42 @@ var _ = ginkgo.Describe("test localstorage volume ", ginkgo.Label("periodCheck")
 				}
 			}
 		})
-		ginkgo.It("Evicte volume", func() {
+		ginkgo.It("update deploy", func() {
+			//delete deploy
 			deployment := &appsv1.Deployment{}
 			deployKey := k8sclient.ObjectKey{
-				Name:      HaDeploymentName,
+				Name:      utils.HaDeploymentName,
 				Namespace: "default",
 			}
 			err := client.Get(ctx, deployKey, deployment)
 			if err != nil {
-				logrus.Printf("%+v ", err)
 				f.ExpectNoError(err)
 			}
 
-			apps, err := labels.NewRequirement("app", selection.In, []string{"demo"})
-			selector := labels.NewSelector()
-			selector = selector.Add(*apps)
-			listOption := k8sclient.ListOptions{
-				LabelSelector: selector,
+			newAffinity := []apiv1.NodeSelectorTerm{
+				{
+					[]apiv1.NodeSelectorRequirement{
+						{
+							Key:      "kubernetes.io/hostname",
+							Operator: apiv1.NodeSelectorOpIn,
+							Values: []string{
+								"k8s-node2",
+							},
+						},
+					},
+					[]apiv1.NodeSelectorRequirement{},
+				},
 			}
-			podlist := &v1.PodList{}
-			err = client.List(ctx, podlist, &listOption)
-			if err != nil {
-				logrus.Printf("%+v ", err)
-				f.ExpectNoError(err)
-			}
+			deployment.Spec.Template.Spec.Affinity.NodeAffinity.RequiredDuringSchedulingIgnoredDuringExecution.NodeSelectorTerms = newAffinity
 
-			logrus.Infof("drain  " + podlist.Items[0].Spec.NodeName)
-			beforeEvictionNode := podlist.Items[0].Spec.NodeName
-			_ = runInLinux("kubectl taint node --all node-role.kubernetes.io/master-")
-			_ = runInLinux("kubectl drain " + podlist.Items[0].Spec.NodeName + " --ignore-daemonsets=true --delete-emptydir-data")
-
-			deployment = &appsv1.Deployment{}
-			deployKey = k8sclient.ObjectKey{
-				Name:      HaDeploymentName,
-				Namespace: "default",
-			}
+			err = client.Update(ctx, deployment)
 			err = client.Get(ctx, deployKey, deployment)
 			if err != nil {
 				logrus.Printf("%+v ", err)
 				f.ExpectNoError(err)
 			}
 			logrus.Infof("waiting for the deployment to be ready ")
-			err = wait.PollImmediate(3*time.Second, 5*time.Minute, func() (done bool, err error) {
+			err = wait.PollImmediate(3*time.Second, 3*time.Minute, func() (done bool, err error) {
 				if err = client.Get(ctx, deployKey, deployment); deployment.Status.AvailableReplicas != int32(1) {
 					return false, nil
 				}
@@ -701,15 +1044,6 @@ var _ = ginkgo.Describe("test localstorage volume ", ginkgo.Label("periodCheck")
 				logrus.Infof("deployment ready timeout")
 				logrus.Error(err)
 			}
-			err = client.List(ctx, podlist, &listOption)
-			if err != nil {
-				logrus.Printf("%+v ", err)
-				f.ExpectNoError(err)
-			}
-
-			logrus.Infof("after drain " + podlist.Items[0].Spec.NodeName)
-			_ = runInLinux("kubectl uncordon " + beforeEvictionNode)
-			gomega.Expect(beforeEvictionNode).NotTo(gomega.Equal(podlist.Items[0].Spec.NodeName))
 		})
 		ginkgo.It("check test file", func() {
 			config, err := config.GetConfig()
@@ -719,7 +1053,7 @@ var _ = ginkgo.Describe("test localstorage volume ", ginkgo.Label("periodCheck")
 
 			deployment := &appsv1.Deployment{}
 			deployKey := k8sclient.ObjectKey{
-				Name:      HaDeploymentName,
+				Name:      utils.HaDeploymentName,
 				Namespace: "default",
 			}
 			err = client.Get(ctx, deployKey, deployment)
@@ -745,7 +1079,7 @@ var _ = ginkgo.Describe("test localstorage volume ", ginkgo.Label("periodCheck")
 			containers := deployment.Spec.Template.Spec.Containers
 			for _, pod := range podlist.Items {
 				for _, container := range containers {
-					output, _, err := ExecInPod(config, deployment.Namespace, pod.Name, "cd /data && cat test", container.Name)
+					output, _, err := utils.ExecInPod(config, deployment.Namespace, pod.Name, "cd /data && cat test", container.Name)
 					if err != nil {
 						logrus.Printf("%+v ", err)
 						f.ExpectNoError(err)
@@ -755,13 +1089,12 @@ var _ = ginkgo.Describe("test localstorage volume ", ginkgo.Label("periodCheck")
 			}
 		})
 	})
-
 	ginkgo.Context("Clean up the environment", func() {
 		ginkgo.It("Delete test Deployment", func() {
 			//delete deploy
 			deployment := &appsv1.Deployment{}
 			deployKey := k8sclient.ObjectKey{
-				Name:      HaDeploymentName,
+				Name:      utils.HaDeploymentName,
 				Namespace: "default",
 			}
 			err := client.Get(ctx, deployKey, deployment)
@@ -787,17 +1120,18 @@ var _ = ginkgo.Describe("test localstorage volume ", ginkgo.Label("periodCheck")
 
 		})
 		ginkgo.It("delete all pvc", func() {
-			err := deleteAllPVC(ctx)
+			err := utils.DeleteAllPVC(ctx)
 			gomega.Expect(err).To(gomega.BeNil())
 		})
 		ginkgo.It("delete all sc", func() {
-			err := deleteAllSC(ctx)
+			err := utils.DeleteAllSC(ctx)
 			gomega.Expect(err).To(gomega.BeNil())
 		})
 		ginkgo.It("delete helm", func() {
-			uninstallHelm()
+			utils.UninstallHelm()
 
 		})
 
 	})
+
 })
