+++ test/e2e/E2eTest/ls_haVolume_test.go
@@ -2,9 +2,9 @@ package E2eTest
 
 import (
 	"context"
+	"github.com/hwameistor/hwameistor/test/e2e/utils"
 	"time"
 
-	clientset "github.com/hwameistor/hwameistor/pkg/apis/client/clientset/versioned/scheme"
 	v1alpha1 "github.com/hwameistor/hwameistor/pkg/apis/hwameistor/v1alpha1"
 	"github.com/hwameistor/hwameistor/test/e2e/framework"
 	"github.com/onsi/ginkgo/v2"
@@ -24,36 +24,28 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client/config"
 )
 
-var _ = ginkgo.Describe("test localstorage volume ", ginkgo.Label("kylin10_arm"), func() {
-	f := framework.NewDefaultFramework(clientset.AddToScheme)
+var _ = ginkgo.Describe("test localstorage Ha volume", ginkgo.Label("periodCheck"), func() {
+
+	f := framework.NewDefaultFramework(v1alpha1.AddToScheme)
 	client := f.GetClient()
 	ctx := context.TODO()
 	ginkgo.It("Configure the base environment", func() {
-		startAdRollback("kylin10arm")
-		f = framework.NewDefaultFramework(clientset.AddToScheme)
-		client = f.GetClient()
-		ctx = context.TODO()
-	})
-
-	ginkgo.It("Configure the base environment", func() {
-		result := configureadEnvironment(ctx, "kylin10arm")
+		result := utils.ConfigureEnvironment(ctx)
 		gomega.Expect(result).To(gomega.BeNil())
-		createLdc(ctx)
-
+		utils.CreateLdc(ctx)
 	})
-
-	ginkgo.Context("create a StorageClass", func() {
+	ginkgo.Context("create a HA-StorageClass", func() {
 		ginkgo.It("create a sc", func() {
 			//create sc
 			deleteObj := apiv1.PersistentVolumeReclaimDelete
 			waitForFirstConsumerObj := storagev1.VolumeBindingWaitForFirstConsumer
 			examplesc := &storagev1.StorageClass{
 				ObjectMeta: metav1.ObjectMeta{
-					Name: "local-storage-hdd-lvm",
+					Name: "local-storage-hdd-lvm-ha",
 				},
 				Provisioner: "lvm.hwameistor.io",
 				Parameters: map[string]string{
-					"replicaNumber":             "1",
+					"replicaNumber":             "2",
 					"poolClass":                 "HDD",
 					"poolType":                  "REGULAR",
 					"volumeKind":                "LVM",
@@ -61,7 +53,7 @@ var _ = ginkgo.Describe("test localstorage volume ", ginkgo.Label("kylin10_arm")
 					"csi.storage.k8s.io/fstype": "xfs",
 				},
 				ReclaimPolicy:        &deleteObj,
-				AllowVolumeExpansion: boolPter(true),
+				AllowVolumeExpansion: utils.BoolPter(true),
 				VolumeBindingMode:    &waitForFirstConsumerObj,
 			}
 			err := client.Create(ctx, examplesc)
@@ -71,13 +63,13 @@ var _ = ginkgo.Describe("test localstorage volume ", ginkgo.Label("kylin10_arm")
 			}
 		})
 	})
-	ginkgo.Context("create a PVC", func() {
+	ginkgo.Context("create a HA-PersistentVolumeClaim", func() {
 		ginkgo.It("create PVC", func() {
 			//create PVC
-			storageClassName := "local-storage-hdd-lvm"
+			storageClassName := "local-storage-hdd-lvm-ha"
 			examplePvc := &apiv1.PersistentVolumeClaim{
 				ObjectMeta: metav1.ObjectMeta{
-					Name:      "pvc-lvm",
+					Name:      "pvc-lvm-ha",
 					Namespace: "default",
 				},
 				Spec: apiv1.PersistentVolumeClaimSpec{
@@ -85,7 +77,7 @@ var _ = ginkgo.Describe("test localstorage volume ", ginkgo.Label("kylin10_arm")
 					StorageClassName: &storageClassName,
 					Resources: apiv1.ResourceRequirements{
 						Requests: apiv1.ResourceList{
-							apiv1.ResourceStorage: resource.MustParse("1Gi"),
+							apiv1.ResourceStorage: resource.MustParse("100Mi"),
 						},
 					},
 				},
@@ -102,15 +94,18 @@ var _ = ginkgo.Describe("test localstorage volume ", ginkgo.Label("kylin10_arm")
 	})
 	ginkgo.Context("create a deployment", func() {
 
-		ginkgo.It("create deployment", func() {
+		ginkgo.It("create a deployment", func() {
 			//create deployment
 			exampleDeployment := &appsv1.Deployment{
 				ObjectMeta: metav1.ObjectMeta{
-					Name:      DeploymentName,
+					Name:      utils.HaDeploymentName,
 					Namespace: "default",
 				},
 				Spec: appsv1.DeploymentSpec{
-					Replicas: int32Ptr(1),
+					Replicas: utils.Int32Ptr(1),
+					Strategy: appsv1.DeploymentStrategy{
+						Type: appsv1.RecreateDeploymentStrategyType,
+					},
 					Selector: &metav1.LabelSelector{
 						MatchLabels: map[string]string{
 							"app": "demo",
@@ -123,6 +118,30 @@ var _ = ginkgo.Describe("test localstorage volume ", ginkgo.Label("kylin10_arm")
 							},
 						},
 						Spec: apiv1.PodSpec{
+							SchedulerName: "hwameistor-scheduler",
+							Affinity: &apiv1.Affinity{
+								NodeAffinity: &apiv1.NodeAffinity{
+									RequiredDuringSchedulingIgnoredDuringExecution: &apiv1.NodeSelector{
+										NodeSelectorTerms: []apiv1.NodeSelectorTerm{
+											{
+												[]apiv1.NodeSelectorRequirement{
+													{
+														Key:      "kubernetes.io/hostname",
+														Operator: apiv1.NodeSelectorOpIn,
+														Values: []string{
+															"k8s-node1",
+														},
+													},
+												},
+												[]apiv1.NodeSelectorRequirement{},
+											},
+										},
+									},
+									PreferredDuringSchedulingIgnoredDuringExecution: nil,
+								},
+								PodAffinity:     nil,
+								PodAntiAffinity: nil,
+							},
 							Containers: []apiv1.Container{
 								{
 									Name:  "web",
@@ -136,7 +155,7 @@ var _ = ginkgo.Describe("test localstorage volume ", ginkgo.Label("kylin10_arm")
 									},
 									VolumeMounts: []apiv1.VolumeMount{
 										{
-											Name:      "2048-volume-lvm",
+											Name:      "2048-volume-lvm-ha",
 											MountPath: "/data",
 										},
 									},
@@ -144,10 +163,10 @@ var _ = ginkgo.Describe("test localstorage volume ", ginkgo.Label("kylin10_arm")
 							},
 							Volumes: []apiv1.Volume{
 								{
-									Name: "2048-volume-lvm",
+									Name: "2048-volume-lvm-ha",
 									VolumeSource: apiv1.VolumeSource{
 										PersistentVolumeClaim: &apiv1.PersistentVolumeClaimVolumeSource{
-											ClaimName: "pvc-lvm",
+											ClaimName: "pvc-lvm-ha",
 										},
 									},
 								},
@@ -161,12 +180,14 @@ var _ = ginkgo.Describe("test localstorage volume ", ginkgo.Label("kylin10_arm")
 				logrus.Printf("%+v ", err)
 				f.ExpectNoError(err)
 			}
+
 			gomega.Expect(err).To(gomega.BeNil())
 		})
 		ginkgo.It("PVC STATUS should be Bound", func() {
+
 			pvc := &apiv1.PersistentVolumeClaim{}
 			pvcKey := k8sclient.ObjectKey{
-				Name:      "pvc-lvm",
+				Name:      "pvc-lvm-ha",
 				Namespace: "default",
 			}
 			err := client.Get(ctx, pvcKey, pvc)
@@ -174,7 +195,6 @@ var _ = ginkgo.Describe("test localstorage volume ", ginkgo.Label("kylin10_arm")
 				logrus.Printf("%+v ", err)
 				f.ExpectNoError(err)
 			}
-
 			logrus.Infof("Waiting for the PVC to be bound")
 			err = wait.PollImmediate(3*time.Second, 3*time.Minute, func() (done bool, err error) {
 				if err = client.Get(ctx, pvcKey, pvc); pvc.Status.Phase != apiv1.ClaimBound {
@@ -191,7 +211,7 @@ var _ = ginkgo.Describe("test localstorage volume ", ginkgo.Label("kylin10_arm")
 		ginkgo.It("deploy STATUS should be AVAILABLE", func() {
 			deployment := &appsv1.Deployment{}
 			deployKey := k8sclient.ObjectKey{
-				Name:      DeploymentName,
+				Name:      utils.HaDeploymentName,
 				Namespace: "default",
 			}
 			err := client.Get(ctx, deployKey, deployment)
@@ -211,31 +231,12 @@ var _ = ginkgo.Describe("test localstorage volume ", ginkgo.Label("kylin10_arm")
 				logrus.Error(err)
 			}
 			gomega.Expect(err).To(gomega.BeNil())
-
 		})
 
 	})
-	ginkgo.Context("Test the volume", func() {
-		ginkgo.It("check lvg", func() {
-			lvrList := &v1alpha1.LocalVolumeReplicaList{}
-			err := client.List(ctx, lvrList)
-			if err != nil {
-				logrus.Printf("list lvr failed ：%+v ", err)
-			}
-			lvgList := &v1alpha1.LocalVolumeGroupList{}
-			err = client.List(ctx, lvgList)
-			if err != nil {
-				logrus.Printf("list lvg failed ：%+v ", err)
-			}
-			for _, lvr := range lvrList.Items {
-				for _, lvg := range lvgList.Items {
-					gomega.Expect(lvr.Spec.NodeName).To(gomega.Equal(lvg.Spec.Accessibility.Nodes[0]))
-				}
-			}
-
-		})
-		ginkgo.It("write test data", func() {
-
+	ginkgo.Context("test volumes", func() {
+		ginkgo.It("Write test file", func() {
+			//create a request
 			config, err := config.GetConfig()
 			if err != nil {
 				return
@@ -243,7 +244,7 @@ var _ = ginkgo.Describe("test localstorage volume ", ginkgo.Label("kylin10_arm")
 
 			deployment := &appsv1.Deployment{}
 			deployKey := k8sclient.ObjectKey{
-				Name:      DeploymentName,
+				Name:      utils.HaDeploymentName,
 				Namespace: "default",
 			}
 			err = client.Get(ctx, deployKey, deployment)
@@ -269,12 +270,12 @@ var _ = ginkgo.Describe("test localstorage volume ", ginkgo.Label("kylin10_arm")
 			containers := deployment.Spec.Template.Spec.Containers
 			for _, pod := range podlist.Items {
 				for _, container := range containers {
-					_, _, err := ExecInPod(config, deployment.Namespace, pod.Name, "cd /data && echo it-is-a-test >test", container.Name)
+					_, _, err := utils.ExecInPod(config, deployment.Namespace, pod.Name, "cd /data && echo it-is-a-test >test", container.Name)
 					if err != nil {
 						logrus.Printf("%+v ", err)
 						f.ExpectNoError(err)
 					}
-					output, _, err := ExecInPod(config, deployment.Namespace, pod.Name, "cd /data && cat test", container.Name)
+					output, _, err := utils.ExecInPod(config, deployment.Namespace, pod.Name, "cd /data && cat test", container.Name)
 					if err != nil {
 						logrus.Printf("%+v ", err)
 						f.ExpectNoError(err)
@@ -283,7 +284,7 @@ var _ = ginkgo.Describe("test localstorage volume ", ginkgo.Label("kylin10_arm")
 				}
 			}
 		})
-		ginkgo.It("Delete test data", func() {
+		ginkgo.It("Delete test file", func() {
 			config, err := config.GetConfig()
 			if err != nil {
 				return
@@ -291,7 +292,7 @@ var _ = ginkgo.Describe("test localstorage volume ", ginkgo.Label("kylin10_arm")
 
 			deployment := &appsv1.Deployment{}
 			deployKey := k8sclient.ObjectKey{
-				Name:      DeploymentName,
+				Name:      utils.HaDeploymentName,
 				Namespace: "default",
 			}
 			err = client.Get(ctx, deployKey, deployment)
@@ -317,12 +318,12 @@ var _ = ginkgo.Describe("test localstorage volume ", ginkgo.Label("kylin10_arm")
 			containers := deployment.Spec.Template.Spec.Containers
 			for _, pod := range podlist.Items {
 				for _, container := range containers {
-					_, _, err := ExecInPod(config, deployment.Namespace, pod.Name, "cd /data && rm -rf test", container.Name)
+					_, _, err := utils.ExecInPod(config, deployment.Namespace, pod.Name, "cd /data && rm -rf test", container.Name)
 					if err != nil {
 						logrus.Printf("%+v ", err)
 						f.ExpectNoError(err)
 					}
-					output, _, err := ExecInPod(config, deployment.Namespace, pod.Name, "cd /data && ls", container.Name)
+					output, _, err := utils.ExecInPod(config, deployment.Namespace, pod.Name, "cd /data && ls", container.Name)
 					if err != nil {
 						logrus.Printf("%+v ", err)
 						f.ExpectNoError(err)
@@ -332,21 +333,204 @@ var _ = ginkgo.Describe("test localstorage volume ", ginkgo.Label("kylin10_arm")
 			}
 		})
 	})
+	ginkgo.Context("test HA-volumes", func() {
+		ginkgo.It("create a localvolumemigrate", func() {
+
+			lvrList := &v1alpha1.LocalVolumeReplicaList{}
+			err := client.List(ctx, lvrList)
+			if err != nil {
+				logrus.Printf("list lvr failed ：%+v ", err)
+			}
+			lvlist := &v1alpha1.LocalVolumeList{}
+			err = client.List(ctx, lvlist)
+			if err != nil {
+				logrus.Error("%+v ", err)
+				f.ExpectNoError(err)
+			}
+
+			lvname := lvlist.Items[0].Name
+			for _, lvr := range lvrList.Items {
+				if lvr.Spec.NodeName == "k8s-master" {
+					exlvm := &v1alpha1.LocalVolumeMigrate{
+						ObjectMeta: metav1.ObjectMeta{
+							Name:      "localvolumegroupmigrate-1",
+							Namespace: "default",
+						},
+						Spec: v1alpha1.LocalVolumeMigrateSpec{
+							TargetNodesSuggested: []string{"k8s-node2"},
+							SourceNode:           "k8s-master",
+							VolumeName:           lvname,
+							MigrateAllVols:       true,
+						},
+					}
+
+					err = client.Create(ctx, exlvm)
+					logrus.Infof("create lvm")
+					if err != nil {
+						logrus.Printf("Create lvgm failed ：%+v ", err)
+						f.ExpectNoError(err)
+					}
+					logrus.Infof("wait 3 minutes for migrate lv")
+					time.Sleep(3 * time.Minute)
+					break
+
+				}
+
+			}
+
+		})
+
+		ginkgo.It("Write test file", func() {
+			config, err := config.GetConfig()
+			if err != nil {
+				return
+			}
+
+			deployment := &appsv1.Deployment{}
+			deployKey := k8sclient.ObjectKey{
+				Name:      utils.HaDeploymentName,
+				Namespace: "default",
+			}
+			err = client.Get(ctx, deployKey, deployment)
+			if err != nil {
+				logrus.Printf("%+v ", err)
+				f.ExpectNoError(err)
+			}
+
+			apps, err := labels.NewRequirement("app", selection.In, []string{"demo"})
+			selector := labels.NewSelector()
+			selector = selector.Add(*apps)
+			listOption := k8sclient.ListOptions{
+				LabelSelector: selector,
+			}
+			podlist := &v1.PodList{}
+			err = client.List(ctx, podlist, &listOption)
+
+			if err != nil {
+				logrus.Printf("%+v ", err)
+				f.ExpectNoError(err)
+			}
+
+			containers := deployment.Spec.Template.Spec.Containers
+			for _, pod := range podlist.Items {
+				for _, container := range containers {
+					_, _, err := utils.ExecInPod(config, deployment.Namespace, pod.Name, "cd /data && echo it-is-a-test >test", container.Name)
+					if err != nil {
+						logrus.Printf("%+v ", err)
+						f.ExpectNoError(err)
+					}
+					output, _, err := utils.ExecInPod(config, deployment.Namespace, pod.Name, "cd /data && cat test", container.Name)
+					if err != nil {
+						logrus.Printf("%+v ", err)
+						f.ExpectNoError(err)
+					}
+					gomega.Expect(output).To(gomega.Equal("it-is-a-test"))
+				}
+			}
+		})
+		ginkgo.It("update deploy", func() {
+			//delete deploy
+			deployment := &appsv1.Deployment{}
+			deployKey := k8sclient.ObjectKey{
+				Name:      utils.HaDeploymentName,
+				Namespace: "default",
+			}
+			err := client.Get(ctx, deployKey, deployment)
+			if err != nil {
+				f.ExpectNoError(err)
+			}
+
+			newAffinity := []apiv1.NodeSelectorTerm{
+				{
+					[]apiv1.NodeSelectorRequirement{
+						{
+							Key:      "kubernetes.io/hostname",
+							Operator: apiv1.NodeSelectorOpIn,
+							Values: []string{
+								"k8s-node2",
+							},
+						},
+					},
+					[]apiv1.NodeSelectorRequirement{},
+				},
+			}
+			deployment.Spec.Template.Spec.Affinity.NodeAffinity.RequiredDuringSchedulingIgnoredDuringExecution.NodeSelectorTerms = newAffinity
+
+			err = client.Update(ctx, deployment)
+			err = client.Get(ctx, deployKey, deployment)
+			if err != nil {
+				logrus.Printf("%+v ", err)
+				f.ExpectNoError(err)
+			}
+			logrus.Infof("waiting for the deployment to be ready ")
+			err = wait.PollImmediate(3*time.Second, 3*time.Minute, func() (done bool, err error) {
+				if err = client.Get(ctx, deployKey, deployment); deployment.Status.AvailableReplicas != int32(1) {
+					return false, nil
+				}
+				return true, nil
+			})
+			if err != nil {
+				logrus.Infof("deployment ready timeout")
+				logrus.Error(err)
+			}
+		})
+		ginkgo.It("check test file", func() {
+			config, err := config.GetConfig()
+			if err != nil {
+				return
+			}
+
+			deployment := &appsv1.Deployment{}
+			deployKey := k8sclient.ObjectKey{
+				Name:      utils.HaDeploymentName,
+				Namespace: "default",
+			}
+			err = client.Get(ctx, deployKey, deployment)
+			if err != nil {
+				logrus.Printf("%+v ", err)
+				f.ExpectNoError(err)
+			}
+
+			apps, err := labels.NewRequirement("app", selection.In, []string{"demo"})
+			selector := labels.NewSelector()
+			selector = selector.Add(*apps)
+			listOption := k8sclient.ListOptions{
+				LabelSelector: selector,
+			}
+			podlist := &v1.PodList{}
+			err = client.List(ctx, podlist, &listOption)
+
+			if err != nil {
+				logrus.Printf("%+v ", err)
+				f.ExpectNoError(err)
+			}
+
+			containers := deployment.Spec.Template.Spec.Containers
+			for _, pod := range podlist.Items {
+				for _, container := range containers {
+					output, _, err := utils.ExecInPod(config, deployment.Namespace, pod.Name, "cd /data && cat test", container.Name)
+					if err != nil {
+						logrus.Printf("%+v ", err)
+						f.ExpectNoError(err)
+					}
+					gomega.Expect(output).To(gomega.Equal("it-is-a-test"))
+				}
+			}
+		})
+	})
 	ginkgo.Context("Clean up the environment", func() {
 		ginkgo.It("Delete test Deployment", func() {
 			//delete deploy
 			deployment := &appsv1.Deployment{}
 			deployKey := k8sclient.ObjectKey{
-				Name:      DeploymentName,
+				Name:      utils.HaDeploymentName,
 				Namespace: "default",
 			}
 			err := client.Get(ctx, deployKey, deployment)
 			if err != nil {
-				logrus.Error(err)
+				logrus.Printf("%+v ", err)
 				f.ExpectNoError(err)
 			}
-			logrus.Infof("deleting test Deployment ")
-
 			err = client.Delete(ctx, deployment)
 			if err != nil {
 				logrus.Error(err)
@@ -364,17 +548,19 @@ var _ = ginkgo.Describe("test localstorage volume ", ginkgo.Label("kylin10_arm")
 			gomega.Expect(err).To(gomega.BeNil())
 
 		})
-		ginkgo.It("delete all pvc ", func() {
-			err := deleteAllPVC(ctx)
+		ginkgo.It("delete all pvc", func() {
+			err := utils.DeleteAllPVC(ctx)
 			gomega.Expect(err).To(gomega.BeNil())
 		})
 		ginkgo.It("delete all sc", func() {
-			err := deleteAllSC(ctx)
+			err := utils.DeleteAllSC(ctx)
 			gomega.Expect(err).To(gomega.BeNil())
 		})
 		ginkgo.It("delete helm", func() {
-			uninstallHelm()
+			utils.UninstallHelm()
+
 		})
+
 	})
 
 })
