+++ test/e2e/E2eTest/performance_test.go
@@ -1,34 +1,45 @@
 package E2eTest
 
 import (
+	"bufio"
 	"context"
+	"github.com/hwameistor/hwameistor/test/e2e/utils"
+	"os"
+	"strings"
 	"time"
 
 	clientset "github.com/hwameistor/hwameistor/pkg/apis/client/clientset/versioned/scheme"
+	v1alpha1 "github.com/hwameistor/hwameistor/pkg/apis/hwameistor/v1alpha1"
 	"github.com/hwameistor/hwameistor/test/e2e/framework"
 	"github.com/onsi/ginkgo/v2"
 	"github.com/onsi/gomega"
 	"github.com/sirupsen/logrus"
 	appsv1 "k8s.io/api/apps/v1"
 	apiv1 "k8s.io/api/core/v1"
+	v1 "k8s.io/api/core/v1"
 	storagev1 "k8s.io/api/storage/v1"
 	k8serror "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/labels"
+	"k8s.io/apimachinery/pkg/selection"
 	"k8s.io/apimachinery/pkg/util/wait"
 	k8sclient "sigs.k8s.io/controller-runtime/pkg/client"
+	"sigs.k8s.io/controller-runtime/pkg/client/config"
 )
 
-var _ = ginkgo.Describe("test fs volume", ginkgo.Label("periodCheck"), func() {
+var _ = ginkgo.Describe("performance testing ", ginkgo.Label("performance-test"), func() {
 
 	f := framework.NewDefaultFramework(clientset.AddToScheme)
 	client := f.GetClient()
 	ctx := context.TODO()
 	ginkgo.It("Configure the base environment", func() {
-		result := configureEnvironment(ctx)
+		result := utils.ConfigureEnvironment(ctx)
 		gomega.Expect(result).To(gomega.BeNil())
+		utils.CreateLdc(ctx)
 
 	})
+
 	ginkgo.Context("create a StorageClass", func() {
 		ginkgo.It("create a sc", func() {
 			//create sc
@@ -36,14 +47,19 @@ var _ = ginkgo.Describe("test fs volume", ginkgo.Label("periodCheck"), func() {
 			waitForFirstConsumerObj := storagev1.VolumeBindingWaitForFirstConsumer
 			examplesc := &storagev1.StorageClass{
 				ObjectMeta: metav1.ObjectMeta{
-					Name: "local-storage-hdd-fs",
+					Name: "local-storage-hdd-lvm",
 				},
-				Provisioner: "disk.hwameistor.io",
+				Provisioner: "lvm.hwameistor.io",
 				Parameters: map[string]string{
-					"diskType": "HDD",
+					"replicaNumber":             "1",
+					"poolClass":                 "HDD",
+					"poolType":                  "REGULAR",
+					"volumeKind":                "LVM",
+					"striped":                   "true",
+					"csi.storage.k8s.io/fstype": "xfs",
 				},
 				ReclaimPolicy:        &deleteObj,
-				AllowVolumeExpansion: boolPter(false),
+				AllowVolumeExpansion: utils.BoolPter(true),
 				VolumeBindingMode:    &waitForFirstConsumerObj,
 			}
 			err := client.Create(ctx, examplesc)
@@ -56,10 +72,10 @@ var _ = ginkgo.Describe("test fs volume", ginkgo.Label("periodCheck"), func() {
 	ginkgo.Context("create a PVC", func() {
 		ginkgo.It("create PVC", func() {
 			//create PVC
-			storageClassName := "local-storage-hdd-fs"
+			storageClassName := "local-storage-hdd-lvm"
 			examplePvc := &apiv1.PersistentVolumeClaim{
 				ObjectMeta: metav1.ObjectMeta{
-					Name:      "pvc-fs",
+					Name:      "pvc-lvm",
 					Namespace: "default",
 				},
 				Spec: apiv1.PersistentVolumeClaimSpec{
@@ -67,7 +83,7 @@ var _ = ginkgo.Describe("test fs volume", ginkgo.Label("periodCheck"), func() {
 					StorageClassName: &storageClassName,
 					Resources: apiv1.ResourceRequirements{
 						Requests: apiv1.ResourceList{
-							apiv1.ResourceStorage: resource.MustParse("1Gi"),
+							apiv1.ResourceStorage: resource.MustParse("2Gi"),
 						},
 					},
 				},
@@ -88,11 +104,11 @@ var _ = ginkgo.Describe("test fs volume", ginkgo.Label("periodCheck"), func() {
 			//create deployment
 			exampleDeployment := &appsv1.Deployment{
 				ObjectMeta: metav1.ObjectMeta{
-					Name:      DeploymentName,
+					Name:      utils.DeploymentName,
 					Namespace: "default",
 				},
 				Spec: appsv1.DeploymentSpec{
-					Replicas: int32Ptr(1),
+					Replicas: utils.Int32Ptr(1),
 					Selector: &metav1.LabelSelector{
 						MatchLabels: map[string]string{
 							"app": "demo",
@@ -109,12 +125,22 @@ var _ = ginkgo.Describe("test fs volume", ginkgo.Label("periodCheck"), func() {
 							Containers: []apiv1.Container{
 								{
 									Name:  "web",
-									Image: "ghcr.m.daocloud.io/daocloud/dao-2048:v1.2.0",
+									Image: "daocloud.io/daocloud/testtools:latest",
 									Ports: []apiv1.ContainerPort{
 										{
 											Name:          "http",
 											Protocol:      apiv1.ProtocolTCP,
 											ContainerPort: 80,
+										},
+									},
+									Resources: apiv1.ResourceRequirements{
+										Limits: apiv1.ResourceList{
+											"cpu":    resource.MustParse("1000m"),
+											"memory": resource.MustParse("4Gi"),
+										},
+										Requests: apiv1.ResourceList{
+											"cpu":    resource.MustParse("1000m"),
+											"memory": resource.MustParse("4Gi"),
 										},
 									},
 									VolumeMounts: []apiv1.VolumeMount{
@@ -130,7 +156,7 @@ var _ = ginkgo.Describe("test fs volume", ginkgo.Label("periodCheck"), func() {
 									Name: "2048-volume-lvm",
 									VolumeSource: apiv1.VolumeSource{
 										PersistentVolumeClaim: &apiv1.PersistentVolumeClaimVolumeSource{
-											ClaimName: "pvc-fs",
+											ClaimName: "pvc-lvm",
 										},
 									},
 								},
@@ -149,7 +175,7 @@ var _ = ginkgo.Describe("test fs volume", ginkgo.Label("periodCheck"), func() {
 		ginkgo.It("PVC STATUS should be Bound", func() {
 			pvc := &apiv1.PersistentVolumeClaim{}
 			pvcKey := k8sclient.ObjectKey{
-				Name:      "pvc-fs",
+				Name:      "pvc-lvm",
 				Namespace: "default",
 			}
 			err := client.Get(ctx, pvcKey, pvc)
@@ -174,7 +200,7 @@ var _ = ginkgo.Describe("test fs volume", ginkgo.Label("periodCheck"), func() {
 		ginkgo.It("deploy STATUS should be AVAILABLE", func() {
 			deployment := &appsv1.Deployment{}
 			deployKey := k8sclient.ObjectKey{
-				Name:      DeploymentName,
+				Name:      utils.DeploymentName,
 				Namespace: "default",
 			}
 			err := client.Get(ctx, deployKey, deployment)
@@ -198,12 +224,105 @@ var _ = ginkgo.Describe("test fs volume", ginkgo.Label("periodCheck"), func() {
 		})
 
 	})
+	ginkgo.Context("Test the volume", func() {
+		ginkgo.It("check lvg", func() {
+			lvrList := &v1alpha1.LocalVolumeReplicaList{}
+			err := client.List(ctx, lvrList)
+			if err != nil {
+				logrus.Printf("list lvr failed ：%+v ", err)
+			}
+			lvgList := &v1alpha1.LocalVolumeGroupList{}
+			err = client.List(ctx, lvgList)
+			if err != nil {
+				logrus.Printf("list lvg failed ：%+v ", err)
+			}
+			for _, lvr := range lvrList.Items {
+				for _, lvg := range lvgList.Items {
+					gomega.Expect(lvr.Spec.NodeName).To(gomega.Equal(lvg.Spec.Accessibility.Nodes[0]))
+				}
+			}
+
+		})
+		ginkgo.It("performance test", func() {
+
+			config, err := config.GetConfig()
+			if err != nil {
+				return
+			}
+
+			deployment := &appsv1.Deployment{}
+			deployKey := k8sclient.ObjectKey{
+				Name:      utils.DeploymentName,
+				Namespace: "default",
+			}
+			err = client.Get(ctx, deployKey, deployment)
+			if err != nil {
+				logrus.Printf("%+v ", err)
+				f.ExpectNoError(err)
+			}
+
+			apps, err := labels.NewRequirement("app", selection.In, []string{"demo"})
+			selector := labels.NewSelector()
+			selector = selector.Add(*apps)
+			listOption := k8sclient.ListOptions{
+				LabelSelector: selector,
+			}
+			podlist := &v1.PodList{}
+			err = client.List(ctx, podlist, &listOption)
+
+			if err != nil {
+				logrus.Printf("%+v ", err)
+				f.ExpectNoError(err)
+			}
+
+			containers := deployment.Spec.Template.Spec.Containers
+			for _, pod := range podlist.Items {
+				for _, container := range containers {
+					output, _, err := utils.ExecInPod(config, deployment.Namespace, pod.Name, "fio -direct=1  -iodepth=128 -rw=randwrite -ioengine=libaio -bs=4K -size=1G -numjobs=1 -runtime=600 -group_reporting -filename=/data/file.txt -name=Rand_Write_IOPS_Test ", container.Name)
+					if err != nil {
+						logrus.Printf("%+v ", err)
+						f.ExpectNoError(err)
+					}
+					filePath1 := "/e2e-test/log/" + time.Now().Format(time.RFC3339) + ".txt"
+					logrus.Printf(filePath1)
+					file, err := os.OpenFile(filePath1, os.O_WRONLY|os.O_CREATE, 0666)
+					if err != nil {
+						logrus.Printf("open file err=%v\n", err)
+						return
+					}
+					defer file.Close()
+					writer := bufio.NewWriter(file)
+					writer.WriteString(output)
+					writer.Flush()
+
+					filePath2 := "/e2e-test/summary.txt"
+					file2, err := os.OpenFile(filePath2, os.O_RDWR|os.O_APPEND|os.O_CREATE, 0666)
+					if err != nil {
+						logrus.Printf("open file err=%v\n", err)
+						return
+					}
+					defer file2.Close()
+					writer2 := bufio.NewWriter(file2)
+					result_head := strings.Index(output, "write: IOPS")
+					result_end := strings.Index(output, "slat")
+					result := time.Now().Format(time.RFC3339) + output[result_head:result_end]
+					if _, err = writer2.WriteString(result); err != nil {
+						logrus.Error(err)
+					}
+					if err = writer2.Flush(); err != nil {
+						logrus.Error(err)
+					}
+					logrus.Info(result)
+				}
+			}
+		})
+	})
 	ginkgo.Context("Clean up the environment", func() {
 		ginkgo.It("Delete test Deployment", func() {
 			//delete deploy
 			deployment := &appsv1.Deployment{}
 			deployKey := k8sclient.ObjectKey{
-				Name:      DeploymentName,
+				Name:      utils.DeploymentName,
 				Namespace: "default",
 			}
 			err := client.Get(ctx, deployKey, deployment)
@@ -231,36 +350,15 @@ var _ = ginkgo.Describe("test fs volume", ginkgo.Label("periodCheck"), func() {
 
 		})
 		ginkgo.It("delete all pvc ", func() {
-			err := deleteAllPVC(ctx)
-			gomega.Expect(err).To(gomega.BeNil())
-		})
-		ginkgo.It("check pv", func() {
-			logrus.Printf("check pv")
-			pvList := &apiv1.PersistentVolumeList{}
-
-			err := wait.PollImmediate(3*time.Second, 3*time.Minute, func() (done bool, err error) {
-				err = client.List(ctx, pvList)
-				if err != nil {
-					logrus.Error("get pv list error", err)
-					f.ExpectNoError(err)
-				}
-				if len(pvList.Items) != 0 {
-					return false, nil
-				} else {
-					return true, nil
-				}
-			})
-			if err != nil {
-				logrus.Error(err)
-			}
+			err := utils.DeleteAllPVC(ctx)
 			gomega.Expect(err).To(gomega.BeNil())
 		})
 		ginkgo.It("delete all sc", func() {
-			err := deleteAllSC(ctx)
+			err := utils.DeleteAllSC(ctx)
 			gomega.Expect(err).To(gomega.BeNil())
 		})
 		ginkgo.It("delete helm", func() {
-			uninstallHelm()
+			utils.UninstallHelm()
 		})
 	})
