+++ test/e2e/adaptation_test/ad_c79_offline_test.go
@@ -1,7 +1,8 @@
-package E2eTest
+package adaptation
 
 import (
 	"context"
+	"github.com/hwameistor/hwameistor/test/e2e/utils"
 	"time"
 
 	clientset "github.com/hwameistor/hwameistor/pkg/apis/client/clientset/versioned/scheme"
@@ -24,21 +25,15 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client/config"
 )
 
-var _ = ginkgo.Describe("test localstorage volume", ginkgo.Label("k8s1.25"), func() {
+var _ = ginkgo.Describe("test localstorage volume", ginkgo.Label("centos7.9_offline"), func() {
+	utils.StartAdRollback("centos7.9_offline")
 	f := framework.NewDefaultFramework(clientset.AddToScheme)
 	client := f.GetClient()
 	ctx := context.TODO()
 	ginkgo.It("Configure the base environment", func() {
-		startAdRollback("k8s1.25")
-		f = framework.NewDefaultFramework(clientset.AddToScheme)
-		client = f.GetClient()
-		ctx = context.TODO()
-	})
-
-	ginkgo.It("Configure the base environment", func() {
-		result := configureadEnvironment(ctx, "k8s1.25")
+		result := utils.ConfigureadEnvironment(ctx, "centos7.9_offline")
 		gomega.Expect(result).To(gomega.BeNil())
-		createLdc(ctx)
+		utils.CreateLdc(ctx)
 
 	})
 
@@ -61,7 +56,7 @@ var _ = ginkgo.Describe("test localstorage volume", ginkgo.Label("k8s1.25"), fun
 					"csi.storage.k8s.io/fstype": "xfs",
 				},
 				ReclaimPolicy:        &deleteObj,
-				AllowVolumeExpansion: boolPter(true),
+				AllowVolumeExpansion: utils.BoolPter(true),
 				VolumeBindingMode:    &waitForFirstConsumerObj,
 			}
 			err := client.Create(ctx, examplesc)
@@ -106,11 +101,11 @@ var _ = ginkgo.Describe("test localstorage volume", ginkgo.Label("k8s1.25"), fun
 			//create deployment
 			exampleDeployment := &appsv1.Deployment{
 				ObjectMeta: metav1.ObjectMeta{
-					Name:      DeploymentName,
+					Name:      utils.DeploymentName,
 					Namespace: "default",
 				},
 				Spec: appsv1.DeploymentSpec{
-					Replicas: int32Ptr(1),
+					Replicas: utils.Int32Ptr(1),
 					Selector: &metav1.LabelSelector{
 						MatchLabels: map[string]string{
 							"app": "demo",
@@ -127,7 +122,7 @@ var _ = ginkgo.Describe("test localstorage volume", ginkgo.Label("k8s1.25"), fun
 							Containers: []apiv1.Container{
 								{
 									Name:  "web",
-									Image: "ghcr.m.daocloud.io/daocloud/dao-2048:v1.2.0",
+									Image: "172.30.45.210/daocloud/dao-2048:v1.2.0",
 									Ports: []apiv1.ContainerPort{
 										{
 											Name:          "http",
@@ -157,10 +152,7 @@ var _ = ginkgo.Describe("test localstorage volume", ginkgo.Label("k8s1.25"), fun
 					},
 				},
 			}
-			logrus.Infof("creating deploy")
-
 			err := client.Create(ctx, exampleDeployment)
-
 			if err != nil {
 				logrus.Printf("%+v ", err)
 				f.ExpectNoError(err)
@@ -195,7 +187,7 @@ var _ = ginkgo.Describe("test localstorage volume", ginkgo.Label("k8s1.25"), fun
 		ginkgo.It("deploy STATUS should be AVAILABLE", func() {
 			deployment := &appsv1.Deployment{}
 			deployKey := k8sclient.ObjectKey{
-				Name:      DeploymentName,
+				Name:      utils.DeploymentName,
 				Namespace: "default",
 			}
 			err := client.Get(ctx, deployKey, deployment)
@@ -247,7 +239,7 @@ var _ = ginkgo.Describe("test localstorage volume", ginkgo.Label("k8s1.25"), fun
 
 			deployment := &appsv1.Deployment{}
 			deployKey := k8sclient.ObjectKey{
-				Name:      DeploymentName,
+				Name:      utils.DeploymentName,
 				Namespace: "default",
 			}
 			err = client.Get(ctx, deployKey, deployment)
@@ -273,12 +265,12 @@ var _ = ginkgo.Describe("test localstorage volume", ginkgo.Label("k8s1.25"), fun
 			containers := deployment.Spec.Template.Spec.Containers
 			for _, pod := range podlist.Items {
 				for _, container := range containers {
-					_, _, err := ExecInPod(config, deployment.Namespace, pod.Name, "cd /data && echo it-is-a-test >test", container.Name)
+					_, _, err := utils.ExecInPod(config, deployment.Namespace, pod.Name, "cd /data && echo it-is-a-test >test", container.Name)
 					if err != nil {
 						logrus.Printf("%+v ", err)
 						f.ExpectNoError(err)
 					}
-					output, _, err := ExecInPod(config, deployment.Namespace, pod.Name, "cd /data && cat test", container.Name)
+					output, _, err := utils.ExecInPod(config, deployment.Namespace, pod.Name, "cd /data && cat test", container.Name)
 					if err != nil {
 						logrus.Printf("%+v ", err)
 						f.ExpectNoError(err)
@@ -295,7 +287,7 @@ var _ = ginkgo.Describe("test localstorage volume", ginkgo.Label("k8s1.25"), fun
 
 			deployment := &appsv1.Deployment{}
 			deployKey := k8sclient.ObjectKey{
-				Name:      DeploymentName,
+				Name:      utils.DeploymentName,
 				Namespace: "default",
 			}
 			err = client.Get(ctx, deployKey, deployment)
@@ -321,12 +313,12 @@ var _ = ginkgo.Describe("test localstorage volume", ginkgo.Label("k8s1.25"), fun
 			containers := deployment.Spec.Template.Spec.Containers
 			for _, pod := range podlist.Items {
 				for _, container := range containers {
-					_, _, err := ExecInPod(config, deployment.Namespace, pod.Name, "cd /data && rm -rf test", container.Name)
+					_, _, err := utils.ExecInPod(config, deployment.Namespace, pod.Name, "cd /data && rm -rf test", container.Name)
 					if err != nil {
 						logrus.Printf("%+v ", err)
 						f.ExpectNoError(err)
 					}
-					output, _, err := ExecInPod(config, deployment.Namespace, pod.Name, "cd /data && ls", container.Name)
+					output, _, err := utils.ExecInPod(config, deployment.Namespace, pod.Name, "cd /data && ls", container.Name)
 					if err != nil {
 						logrus.Printf("%+v ", err)
 						f.ExpectNoError(err)
@@ -341,7 +333,7 @@ var _ = ginkgo.Describe("test localstorage volume", ginkgo.Label("k8s1.25"), fun
 			//delete deploy
 			deployment := &appsv1.Deployment{}
 			deployKey := k8sclient.ObjectKey{
-				Name:      DeploymentName,
+				Name:      utils.DeploymentName,
 				Namespace: "default",
 			}
 			err := client.Get(ctx, deployKey, deployment)
@@ -369,15 +361,15 @@ var _ = ginkgo.Describe("test localstorage volume", ginkgo.Label("k8s1.25"), fun
 
 		})
 		ginkgo.It("delete all pvc ", func() {
-			err := deleteAllPVC(ctx)
+			err := utils.DeleteAllPVC(ctx)
 			gomega.Expect(err).To(gomega.BeNil())
 		})
 		ginkgo.It("delete all sc", func() {
-			err := deleteAllSC(ctx)
+			err := utils.DeleteAllSC(ctx)
 			gomega.Expect(err).To(gomega.BeNil())
 		})
 		ginkgo.It("delete helm", func() {
-			uninstallHelm()
+			utils.UninstallHelm()
 		})
 	})
