+++ test/e2e/e2e_test.go
@@ -2,6 +2,7 @@ package e2e
 
 import (
 	_ "github.com/hwameistor/hwameistor/test/e2e/E2eTest"
+	_ "github.com/hwameistor/hwameistor/test/e2e/adaptation_test"
 	"github.com/onsi/ginkgo/v2"
 	"github.com/onsi/gomega"
 	"math/rand"
